Step 2: move the message interface into the "meeseeks" package
