+++ meeseeks/request/request.go
@@ -3,7 +3,7 @@ package request
 import (
 	"errors"
 
-	"github.com/gomeeseeks/meeseeks-box/meeseeks/message"
+	"github.com/gomeeseeks/meeseeks-box/meeseeks"
 	"github.com/gomeeseeks/meeseeks-box/meeseeks/request/parser"
 	"github.com/sirupsen/logrus"
 )
@@ -25,7 +25,7 @@ type Request struct {
 }
 
 // FromMessage gets a message and generates a valid request from it
-func FromMessage(msg message.Message) (Request, error) {
+func FromMessage(msg meeseeks.Message) (Request, error) {
 	args, err := parser.Parse(msg.GetText())
 	logrus.Debugf("Command %s parsed as %s", msg.GetText(), args)
